# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "篮球课堂训练复盘",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/analyses/upload": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Upload Analysis",
        "operationId": "upload_analysis_api_v1_analyses_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_analysis_api_v1_analyses_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/preset": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Rerun Preset",
        "operationId": "rerun_preset_api_v1_analyses_preset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetRerunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Get Analysis",
        "operationId": "get_analysis_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "analyses"
        ],
        "summary": "Delete Analysis",
        "operationId": "delete_analysis_api_v1_analyses__analysis_id__delete",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/result": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Get Result",
        "operationId": "get_result_api_v1_analyses__analysis_id__result_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/media/{kind}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Analysis Media",
        "operationId": "analysis_media_api_v1_analyses__analysis_id__media__kind__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "video/mp4": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/cancel": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Cancel Analysis",
        "operationId": "cancel_analysis_api_v1_analyses__analysis_id__cancel_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/retry": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Retry Analysis",
        "operationId": "retry_analysis_api_v1_analyses__analysis_id__retry_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisPublic": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "analyst_locale": {
            "type": "string",
            "enum": [
              "zh",
              "en"
            ],
            "title": "Analyst Locale",
            "default": "zh"
          },
          "enrollment_mode": {
            "type": "string",
            "enum": [
              "sequential",
              "lineup"
            ],
            "title": "Enrollment Mode",
            "default": "sequential"
          },
          "expected_persons": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Persons"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "stage_message": {
            "type": "string",
            "title": "Stage Message"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "created_via": {
            "type": "string",
            "title": "Created Via"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "mode",
          "source_type",
          "preset_id",
          "status",
          "progress",
          "stage_message",
          "error_code",
          "error_message",
          "owner_id",
          "submitted_at",
          "created_via",
          "retry_count",
          "created_at",
          "updated_at",
          "started_at",
          "completed_at"
        ],
        "title": "AnalysisPublic"
      },
...
      "Body_upload_analysis_api_v1_analyses_upload_post": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "mode": {
            "type": "string",
            "enum": [
              "quick",
              "full"
            ],
            "title": "Mode"
          },
          "enrollment_mode": {
            "type": "string",
            "enum": [
              "sequential",
              "lineup"
            ],
            "title": "Enrollment Mode"
          },
          "expected_persons": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Expected Persons"
          },
          "analyst_locale": {
            "type": "string",
            "enum": [
              "zh",
              "en"
            ],
            "title": "Analyst Locale"
          },
          "sync": {
            "type": "string",
            "maxLength": 8192,
            "title": "Sync"
          },
          "enrollment_video": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Enrollment Video"
          },
          "cam_01": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Cam 01"
          },
          "cam_02": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Cam 02"
          },
          "cam_03": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Cam 03"
          },
          "cam_04": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Cam 04"
          }
        },
        "type": "object",
        "required": [
          "title",
          "mode",
          "enrollment_mode",
          "expected_persons",
          "analyst_locale",
          "sync",
          "enrollment_video",
          "cam_01",
          "cam_02",
          "cam_03",
          "cam_04"
        ],
        "title": "Body_upload_analysis_api_v1_analyses_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PresetRerunRequest": {
        "properties": {
          "preset_id": {
            "type": "string",
            "title": "Preset Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "quick",
              "full"
            ],
            "title": "Mode",
            "default": "full"
          },
          "analyst_locale": {
            "type": "string",
            "enum": [
              "zh",
              "en"
            ],
            "title": "Analyst Locale",
            "default": "zh"
          }
        },
        "type": "object",
        "required": [
          "preset_id"
        ],
        "title": "PresetRerunRequest"
      },
...
      "ProductResult": {
        "properties": {
          "registered_participant_count": {
            "type": "integer",
            "title": "Registered Participant Count"
          },
          "action_counts": {
            "$ref": "#/components/schemas/ProductActionCounts"
          },
          "unsupported_event_count": {
            "type": "integer",
            "title": "Unsupported Event Count"
          },
          "shots": {
            "$ref": "#/components/schemas/ProductShotSummary"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ProductActionEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "media": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Media"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "AI 识别结果，仅供训练复盘。"
          }
        },
        "type": "object",
        "required": [
          "registered_participant_count",
          "action_counts",
          "unsupported_event_count",
          "shots",
          "events",
          "media"
        ],
        "title": "ProductResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}